+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/dataresources/offers/OfferOverview.groovy
@@ -1,6 +1,8 @@
 package life.qbic.portal.offermanager.dataresources.offers
 
+import groovy.transform.EqualsAndHashCode
 import life.qbic.datamodel.dtos.business.OfferId
+import life.qbic.datamodel.dtos.projectmanagement.ProjectIdentifier
 
 /**
  * This class holds data for an offer overview
@@ -11,10 +13,15 @@ import life.qbic.datamodel.dtos.business.OfferId
  *
  * @since 1.0.0
  */
+@EqualsAndHashCode
 class OfferOverview {
 
     final String projectTitle
 
+    /**
+     * @deprecated Use the {@link #associatedProject} property to link an offer with a project
+     */
+    @Deprecated
     final String projectId
 
     final String customer
@@ -25,6 +32,9 @@ class OfferOverview {
 
     final OfferId offerId
 
+    final Optional<ProjectIdentifier> associatedProject
+
+    @Deprecated
     OfferOverview(
             OfferId offerId,
             Date modificationDate,
@@ -38,5 +48,22 @@ class OfferOverview {
         this.projectTitle = projectTitle
         this.customer = customer
         this.totalPrice = totalPrice
+        this.associatedProject = Optional.empty()
+    }
+
+    OfferOverview(
+            OfferId offerId,
+            Date modificationDate,
+            String projectTitle,
+            String customer,
+            double totalPrice,
+            ProjectIdentifier associatedProject) {
+        this.offerId = offerId
+        this.modificationDate = modificationDate
+        this.projectId = ""
+        this.projectTitle = projectTitle
+        this.customer = customer
+        this.totalPrice = totalPrice
+        this.associatedProject = Optional.of(associatedProject)
     }
 }